init erro.py

+++ lib/dct.h
@@ -10,6 +10,6 @@
 
 void init_cos_lut();
 void dct(float energies[], int num_filters, float ceps[NUM_CEPS]);
-void dct_fixed(int8_t energies_q15[], int num_filters, int16_t ceps_q15[NUM_CEPS]);
+void dct_fixed(int32_t energies_q15[], int num_filters, int32_t ceps_q15[NUM_CEPS]);
 
 #endif // !__DCT_H__
